Improve handling of style variable in template.
The `style` variable is missing in Sphinx 7. Instead, it now uses a list
of `styles` [1]. This new list was already added in Sphinx 5.1. It looks
like the `style` was only removed in Sphinx 7, though.

This commit updates the `layout.html` template to define the styles
variable if it is missing. This makes the change compatible with both
Sphinx <5.1 and versions later than that.

[1] https://www.sphinx-doc.org/en/master/development/templating.html#styles